nuxt_django/backend/requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
backports.csv==1.0.7
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
cssselect==1.0.3
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.2
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.0
django-import-export==1.2.0
django-reversion==3.0.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.5
et-xmlfile==1.0.1
fake-useragent==0.1.11
future==0.17.1
httplib2==0.9.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.1
jsonpath==0.81
keyring==10.6.0
keyrings.alt==3.0
lxml==4.3.2
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
openpyxl==2.6.2
parsel==1.5.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycrypto==2.6.1
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.7.2
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pytz==2018.9
pyxdg==0.25
PyYAML==5.1
queuelib==1.5.0
redis==3.2.1
requests==2.21.0
Scrapy==1.6.0
scrapy-fake-useragent==1.1.0
scrapy-redis==0.6.8
scrapyd==1.2.0
scrapyd-client==1.1.0
SecretStorage==2.3.1
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
Twisted==18.9.0
uritemplate==3.0.0
urllib3==1.24.1
w3lib==1.20.0
xlrd==1.2.0
XlsxWriter==1.1.6
xlwt==1.3.0
zope.interface==4.6.0
python-scrapyd-api==2.1.2
